+++ src/test/java/subway/line/application/StationServiceTest.java
@@ -1,4 +1,4 @@
-package subway.application;
+package subway.line.application;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
