+++ src/test/java/subway/line/domain/UpstreamTerminalSectionTest.java
@@ -1,11 +1,7 @@
-package subway.domain;
+package subway.line.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import subway.line.domain.AbstractSection;
-import subway.line.domain.DownstreamTerminalSection;
-import subway.line.domain.MiddleSection;
-import subway.line.domain.UpstreamTerminalSection;
 import subway.line.exception.DuplicateStationInLineException;
 import subway.station.domain.DummyTerminalStation;
