+++ src/test/java/subway/station/repository/StationRepositoryTest.java
@@ -1,4 +1,4 @@
-package subway.repository;
+package subway.station.repository;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import subway.station.dao.StationDao;
-import subway.station.repository.StationRepository;
 
 import java.util.Optional;
